+++ CHANGELOG.md
@@ -38,6 +38,7 @@ BUG FIXES:
   - guests/windows: Reboot after hostname change. [GH-3987]
   - hosts/arch: NFS works with latest versions. [GH-4224]
   - hosts/windows: RDP command works without crash. [GH-3962]
+  - providers/docker: Port on its own will choose random host port. [GH-3991]
   - providers/virtualbox: Increase network device limit to 36. [GH-4206]
   - provisioners/chef: Chef client cleanup should work. [GH-4099]
   - provisioners/puppet: Manifest file can be a directory. [GH-4169]